fix(render): figlet is an ascii_art style, not its own renderer
Backed out the standalone \`figlet\` renderer added in the previous
commit. Earlier in the design discussion I'd agreed that figlet
belongs as a style variant of \`ascii_art\` (same goal — letters
made visible — just a different visual algorithm), but then I
went and made it a peer renderer anyway. Fixing.

\`ascii_art\` now branches on \`style\`:

    render = "ascii_art"                                      # default = blocks
    render = { type = "ascii_art", style = "blocks" }         # tui-big-text blocks
    render = { type = "ascii_art", style = "blocks",
               pixel_size = "quadrant" }                      # + sub-option
    render = { type = "ascii_art", style = "figlet" }         # figlet-rs

\`RenderOptions\` picked up a \`style: Option<String>\`; \`pixel_size\`
stays and is now documented as blocks-specific. Future styles
(custom figlet fonts, animated) hang off this same renderer with
their own sub-options.

The \`figlet\` test file is deleted and its coverage moved into
\`ascii_art::tests\` as \`figlet_style_routes_through_figlet_rs\`.
